requirements.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
black==21.5b1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cryptography==3.4.8
decorator==5.1.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.7
django-allauth==0.45.0
django-cors-headers==3.10.0
django-pydenticon==0.2
django-taggit==2.0.0
django-taggit-serializer==0.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
Faker==10.0.0
flake8==3.9.2
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
ipython==7.29.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mixer==7.2.0
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.7.0
pycparser==2.20
pydenticon==0.3.1
pyflakes==2.3.1
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-xdist-2.5.0
python-dateutil==2.8.2
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
traitlets==5.1.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
